+++ sperl_package.h
@@ -10,7 +10,7 @@ struct SPerl_package {
   SPerl_ARRAY* op_subs;
   int64_t byte_size;
   int64_t constant_pool_address;
-  int64_t name_constant_pool_address;
+  size_t name_constant_pool_address;
 };
 
 SPerl_PACKAGE* SPerl_PACKAGE_new(SPerl* sperl);
